refs #965 Init Content presenters
