+++ .travis.yml
@@ -1,11 +1,5 @@
 language: php
 
-services:
-  - xvfb
-
-addons:
-  chrome: stable
-
 php:
   - 7.2
   - 7.3
@@ -24,11 +18,11 @@ before_script:
    - travis_retry composer self-update
    - travis_retry composer clear-cache
    - travis_retry composer install --no-interaction --prefer-dist --no-suggest
+   - mkdir vendor/orchestra/testbench-core/laravel/storage/app/public
    - chmod -R o+w vendor/orchestra
-   - export DISPLAY=:99.0
+   - vendor/orchestra/testbench-dusk/create-sqlite-db
    - vendor/bin/dusk-updater update
    - vendor/laravel/dusk/bin/chromedriver-linux &
-   - vendor/orchestra/testbench-dusk/create-sqlite-db
    - google-chrome-stable --headless --disable-gpu --remote-debugging-port=9222 http://localhost &
 
 script:
