+++ install-stubs/Orchid/Presenters/UserPresenter.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace App\Orchid\Presenters;
 
 use Orchid\Screen\Presenter;
-use Orchid\Screen\Presenters\Search;
-use Orchid\Screen\Presenters\User;
+use Orchid\Screen\Presenters\Profilable;
+use Orchid\Screen\Presenters\Searchable;
 
-class UserPresenter extends Presenter implements Search, User
+class UserPresenter extends Presenter implements Searchable, Profilable
 {
     /**
      * @return string
