+++ resources/views/contents/profile.blade.php
